// src/app/models/member-register-dto.ts
export class MemberRegisterDto {
    memberId: number;
    name: string;
    surname: string;
    memberNumber: string;
    tcnumber: string;
    statusParameterId: number;
    associationBranchParameterId: number;
    memberDate: string;
    birthOfDate: string;
    
    birthCityId: number;
    decisionNumber: string;
    decisionDate: string;
    isDisabled: number;
    phoneNumber: string;
    
    motherName: string;
    fatherName: string;
}